Transaction Modelv2

+++ Models/Transaction.cs
@@ -3,7 +3,7 @@ namespace CrimsonClothing_api.Models
     public class Transaction
     {
         public int transactionId {get; set;}
-        public int profit {get; set;}
+        public decimal profit {get; set;}
         public int customerId {get; set;}
         public int itemId{get; set;}
